+++ api/src/main/java/io/trenchcrusade/api/warband/deed/Deed.java
@@ -6,8 +6,8 @@ import jakarta.persistence.*;
 public class Deed {
     @Id // KEY
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
-    public Integer getId() {
+    private Long id;
+    public Long getId() {
         return id;
     }
